betting_bot/requirements.txt
python-telegram-bot[job-queue]==22.6
python-dotenv==1.0.0